created utils.py to hold utility functions. moved parse_url() to it, created decode_file_content() to handle decoding content from responses to api requests
